[#1108] Created _class8_plot_coverage_broadband.py
